requirements.txt
pymysql==1.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
aiohttp==3.9.3
